[refactor] ProjectPostAuditService를 인터페이스로 분리
